🔧 refactor: Reorganize data type imports and bump version
Move type definitions to a dedicated module and update
version across project files. Improves code organization
and maintains version consistency.
