+++ core/pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "Stocksense-Core"
-version = "0.5.6"
+version = "0.5.7"
 description = "Add your description here"
 readme = "README.md"
 authors = [{ name = "Akash Desarda", email = "desardaakash@gmail.com" }]
