+++ core/src/stocksense/data/types.py
